Alterações na configuração do swagger para adicionar mais informações da API.
