+++ src/main/java/br/com/guilherme/gastos/GastosApplication.java
@@ -6,27 +6,20 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
 import springfox.documentation.builders.PathSelectors;
 import springfox.documentation.builders.RequestHandlerSelectors;
+import springfox.documentation.service.ApiInfo;
 import springfox.documentation.spi.DocumentationType;
 import springfox.documentation.spring.web.plugins.Docket;
 import springfox.documentation.swagger2.annotations.EnableSwagger2;
 
+import java.util.ArrayList;
+
 @SpringBootApplication
 @EnableEncryptableProperties
-@EnableSwagger2
 public class GastosApplication {
 
     public static void main(String[] args) {
 
         SpringApplication.run(GastosApplication.class, args);
-    }
-
-    @Bean
-    public Docket api(){
-        return new Docket(DocumentationType.SPRING_WEB)
-                        .select()
-                        .apis(RequestHandlerSelectors.basePackage("br.com.guilherme.gastos.controller"))
-                        .paths(PathSelectors.any())
-                        .build();
     }
 
 }
